chore(ui): Move tailwind config from packages into platform folder
