+++ ui/apps/platform/src/css/light.theme.css
