+++ ui/lerna.json
@@ -1,5 +1,5 @@
 {
-    "packages": ["packages/tailwind-config", "apps/platform"],
+    "packages": ["apps/platform"],
     "version": "independent",
     "npmClient": "yarn",
     "useWorkspaces": true,
