--- ui/packages/README.md
@@ -1,5 +1,0 @@
-# StackRox Platform Packages
-
-Please refer to the parent [README.md](../README.md) to learn more about the
-monorepo directory structure and the nature of NPM packages under this
-directory.
