--- ui/packages/tailwind-config/.gitignore
@@ -1,2 +1,0 @@
-# build
-/tailwind.css
