--- ui/packages/tailwind-config/package.json
@@ -1,39 +1,0 @@
-{
-    "name": "@stackrox/tailwind-config",
-    "version": "0.2.0",
-    "description": "StackRox base Tailwind config and CSS themes",
-    "main": "tailwind.config.js",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/stackrox/rox.git",
-        "directory": "ui/packages/tailwind-config"
-    },
-    "license": "UNLICENSED",
-    "files": [
-        "plugins",
-        "tailwind.config.js",
-        "light.theme.css",
-        "dark.theme.css",
-        "tailwind.css"
-    ],
-    "scripts": {
-        "clean": "rm -f tailwind.css",
-        "build": "postcss ./index.tw.css -o ./tailwind.css",
-        "start": "yarn build",
-        "lint": "echo lint",
-        "prepublishOnly": "yarn clean && yarn build"
-    },
-    "dependencies": {
-        "@tailwindcss/forms": "^0.2.1"
-    },
-    "devDependencies": {
-        "autoprefixer": "^10.4.5",
-        "eslint": "^7.32.0",
-        "eslint-config-prettier": "^7.2.0",
-        "eslint-plugin-prettier": "^4.0.0",
-        "postcss": "^8.3.4",
-        "postcss-cli": "^8.3.1",
-        "prettier": "^2.8.7",
-        "tailwindcss": "^2.0.3"
-    }
-}
