--- ui/packages/tailwind-config/postcss.config.js
@@ -1,5 +1,0 @@
-const tailwindcss = require('tailwindcss');
-
-module.exports = {
-    plugins: [tailwindcss('./tailwind.config.js')],
-};
